pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "autodismediasend"
version = "0.1.0"
description = "Send paired MP4+GIF media to Discord via CLI/GUI"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [{ name = "AutoDisMediaSend" }]
dependencies = [
    "requests>=2.31",
    "python-dotenv>=1.0",
    "typer>=0.9",
    "rich>=13.0",
    "platformdirs>=3.0",
]

[project.urls]
Homepage = "https://github.com/0-FoxHunt-0/disdrop"

[project.scripts]
discord-send = "src.cli:app"

[tool.setuptools]
package-dir = {"" = "."}
packages = ["src"]
include-package-data = true

